Deprecate Old Playstate, not entirely removing it
